Refactor: Merge Schema Logic to SQL Base
Move all complicated schema merging logic to SQL Base. Aim to make interface for implementing class as simple as a handful of SQL statements.
