+++ target_postgres/postgres.py
@@ -1,3 +1,4 @@
+from copy import deepcopy
 import io
 import re
 import csv
@@ -225,21 +226,26 @@ class PostgresTarget(SQLInterface):
         self._validate_identifier(new_idenfitier)
         return new_idenfitier
 
-    def update_table_schema(self, cur, remote_table_json_schema, table_json_schema, metadata):
+    def upsert_table(self, cur, remote_table_json_schema, table_json_schema, metadata):
         self._validate_identifier(table_json_schema['name'])
 
-        if remote_table_json_schema is not None:
-            self.merge_put_schemas(cur,
-                                   remote_table_json_schema['name'],
-                                   remote_table_json_schema,
-                                   table_json_schema)
-        else:
-            self.create_table(cur,
-                              table_json_schema['name'],
-                              table_json_schema,
-                              metadata.get('version', None))
+        if remote_table_json_schema is None:
+            create_table_sql = sql.SQL('CREATE TABLE {}.{}').format(
+                sql.Identifier(self.postgres_schema),
+                sql.Identifier(table_json_schema['name']))
 
-        return self.get_table_schema(cur, table_json_schema['name'])
+            cur.execute(sql.SQL('{} ();').format(create_table_sql))
+
+            if 'key_properties' in table_json_schema:
+                self.set_table_metadata(cur, table_json_schema['name'],
+                                        {'key_properties': table_json_schema['key_properties'],
+                                         'version': metadata.get('version', None)})
+
+        return self.upsert_table_helper(cur,
+                                        table_json_schema['name'],
+                                        table_json_schema,
+                                        metadata)
+
 
     def get_update_sql(self, target_table_name, temp_table_name, key_properties, subkeys):
         full_table_name = sql.SQL('{}.{}').format(
@@ -375,10 +381,12 @@ class PostgresTarget(SQLInterface):
         target_table_name = self.get_temp_table_name()
 
         ## Create temp table to upload new data to
-        self.create_table(cur,
-                          target_table_name,
-                          remote_schema,
-                          remote_schema['version'])
+        target_schema = deepcopy(remote_schema)
+        target_schema['name'] = target_table_name
+        self.upsert_table(cur,
+                          None,
+                          target_schema,
+                          {'version': remote_schema['version']})
 
         ## Make streamable CSV records
         csv_headers = list(remote_schema['schema']['properties'].keys())
@@ -488,28 +496,6 @@ class PostgresTarget(SQLInterface):
             sql.Identifier(table_name),
             sql.Literal(json.dumps(parsed_metadata))))
 
-
-    def create_table(self, cur, table_name, schema, table_version):
-        self._validate_identifier(table_name)
-
-        create_table_sql = sql.SQL('CREATE TABLE {}.{}').format(
-                sql.Identifier(self.postgres_schema),
-                sql.Identifier(table_name))
-
-        cur.execute(sql.SQL('{} ();').format(create_table_sql))
-
-        if 'key_properties' in schema:
-            self.set_table_metadata(cur, table_name,
-                                    {'key_properties': schema['key_properties'],
-                                     'version': table_version})
-
-        remote_table_json_schema = self.get_table_schema(cur, table_name)
-
-        self.merge_put_schemas(cur,
-                               remote_table_json_schema['name'],
-                               remote_table_json_schema,
-                               schema)
-
     def get_temp_table_name(self):
         return 'tmp_' + str(uuid.uuid4()).replace('-', '_')
 
@@ -547,7 +533,7 @@ class PostgresTarget(SQLInterface):
         self.set_table_metadata(cur, table_name, metadata)
 
 
-    def remove_column_mapping(self, cur, table_name, mapped_name):
+    def drop_column_mapping(self, cur, table_name, mapped_name):
         metadata = self.get_table_metadata(cur, table_name)
 
         if not metadata:
@@ -590,215 +576,3 @@ class PostgresTarget(SQLInterface):
         metadata['schema'] = {'properties': properties}
 
         return metadata
-
-    def mapping_name(self, field, schema):
-        return field + SEPARATOR + json_schema.sql_shorthand(schema)
-
-    def get_mapping(self, existing_schema, field, schema):
-        canonicalized_field = self.canonicalize_identifier(field)
-        typed_field = self.mapping_name(canonicalized_field, schema)
-
-        if 'mappings' not in existing_schema:
-            return None
-
-        if canonicalized_field in existing_schema['mappings'] \
-                and field == existing_schema['mappings'][canonicalized_field]['from'] \
-                and json_schema.get_type(json_schema.make_nullable(schema)) \
-                == json_schema.get_type(json_schema.make_nullable(existing_schema['mappings'][canonicalized_field])):
-
-            return canonicalized_field
-
-        if typed_field in existing_schema['mappings'] \
-                and field == existing_schema['mappings'][typed_field]['from']:
-
-            return typed_field
-
-        return None
-
-    def split_column(self, cur, table_name, raw_column_name, column_schema, existing_schema):
-        existing_properties = existing_schema['schema']['properties']
-        column_name = self.canonicalize_identifier(raw_column_name)
-
-        if self.get_mapping(existing_schema, raw_column_name, existing_properties[column_name]):
-            self.remove_column_mapping(cur, table_name, column_name)
-
-        ## column_name -> column_name__<current-type>, column_name__<new-type>
-        existing_column_mapping = self.mapping_name(column_name, existing_properties[column_name])
-        new_column_mapping = self.mapping_name(column_name, column_schema)
-
-        ## Update existing properties
-        existing_properties[existing_column_mapping] = json_schema.make_nullable(existing_properties[column_name])
-        existing_properties[new_column_mapping] = json_schema.make_nullable(column_schema)
-
-        ## Add new columns
-        ### NOTE: all migrated columns will be nullable and remain that way
-
-        #### Table Metadata
-        self.add_column_mapping(cur, table_name, raw_column_name,
-                                existing_column_mapping,
-                                existing_properties[existing_column_mapping])
-        self.add_column_mapping(cur, table_name, raw_column_name,
-                                new_column_mapping,
-                                existing_properties[new_column_mapping])
-
-        #### Columns
-        self.add_column(cur,
-                        table_name,
-                        existing_column_mapping,
-                        existing_properties[existing_column_mapping])
-
-        self.add_column(cur,
-                        table_name,
-                        new_column_mapping,
-                        existing_properties[new_column_mapping])
-
-        ## Migrate existing data
-        self.migrate_column(cur,
-                            table_name,
-                            column_name,
-                            existing_column_mapping)
-
-        ## Drop existing column
-        self.drop_column(cur,
-                         table_name,
-                         column_name)
-
-        ## Remove column (field) from existing_properties
-        del existing_properties[column_name]
-
-    def merge_put_schemas(self, cur, table_name, existing_schema, new_schema):
-        new_columns = new_schema['schema']['properties']
-        existing_columns = existing_schema['schema']['properties']
-        existing_columns_raw_names = [v['from'] for v in existing_schema.get('mappings', {}).values()]
-        table_empty = self.is_table_empty(cur, table_name)
-
-        for raw_column_name, column_schema in new_columns.items():
-            canonicalized_column_name = self.canonicalize_identifier(raw_column_name)
-            canonicalized_typed_column_name = self.mapping_name(canonicalized_column_name, column_schema)
-            nullable_column_schema = json_schema.make_nullable(column_schema)
-
-            ## NAME COLLISION
-            if raw_column_name != canonicalized_column_name \
-                    and raw_column_name not in existing_columns_raw_names \
-                    and (canonicalized_column_name in existing_columns
-                         or canonicalized_typed_column_name in existing_columns):
-                raise PostgresError(
-                    'NAME COLLISION: Cannot handle merging column `{}` (canonicalized as: `{}`, canonicalized with type as: `{}`) in table `{}`.'.format(
-                        self.postgres_schema,
-                        raw_column_name,
-                        canonicalized_column_name,
-                        canonicalized_typed_column_name,
-                        table_name
-                    ))
-
-
-            ## EXISTING COLUMNS
-            elif canonicalized_column_name in existing_columns \
-                    and json_schema.to_sql(column_schema) \
-                    == json_schema.to_sql(existing_columns[canonicalized_column_name]):
-                pass
-            ###
-            elif canonicalized_typed_column_name in existing_columns \
-                    and json_schema.to_sql(column_schema) \
-                    == json_schema.to_sql(existing_columns[canonicalized_typed_column_name]):
-                pass
-            ###
-            elif canonicalized_column_name in existing_columns \
-                    and json_schema.to_sql(nullable_column_schema) \
-                    == json_schema.to_sql(existing_columns[canonicalized_column_name]):
-                pass
-            ###
-            elif canonicalized_typed_column_name in existing_columns \
-                    and json_schema.to_sql(nullable_column_schema) \
-                    == json_schema.to_sql(existing_columns[canonicalized_typed_column_name]):
-                pass
-
-            ## NULL COMPATIBILITY
-            elif canonicalized_column_name in existing_columns \
-                    and json_schema.to_sql(nullable_column_schema) == json_schema.to_sql(
-                json_schema.make_nullable(existing_columns[canonicalized_column_name])):
-
-                ## MAKE NULLABLE
-                self.make_column_nullable(cur,
-                                          table_name,
-                                          canonicalized_column_name)
-                existing_columns[canonicalized_column_name] = json_schema.make_nullable(
-                    existing_columns[canonicalized_column_name])
-
-            ## FIRST DUPLICATE TYPE
-            elif canonicalized_column_name in existing_columns:
-
-                self.split_column(cur,
-                                  table_name,
-                                  raw_column_name,
-                                  column_schema,
-                                  existing_schema)
-
-            ## MULTI DUPLICATE TYPE
-            elif raw_column_name in existing_columns_raw_names:
-
-                ## Add new column
-                self.add_column_mapping(cur, table_name, raw_column_name,
-                                        canonicalized_typed_column_name,
-                                        nullable_column_schema)
-                existing_columns_raw_names.append(canonicalized_typed_column_name)
-
-                self.add_column(cur,
-                                table_name,
-                                canonicalized_typed_column_name,
-                                nullable_column_schema)
-
-                ## Update existing properties
-                existing_columns[canonicalized_typed_column_name] = nullable_column_schema
-
-            ## NEW COLUMN, VALID NAME, EMPTY TABLE
-            elif canonicalized_column_name == raw_column_name and table_empty:
-
-                self.add_column(cur,
-                                table_name,
-                                canonicalized_column_name,
-                                column_schema)
-                existing_columns[canonicalized_column_name] = column_schema
-
-            ## NEW COLUMN, VALID NAME
-            elif canonicalized_column_name == raw_column_name:
-
-                self.add_column(cur,
-                                table_name,
-                                canonicalized_column_name,
-                                nullable_column_schema)
-                existing_columns[canonicalized_column_name] = nullable_column_schema
-
-            ## NEW COLUMN, INVALID NAME, EMPTY TABLE
-            elif canonicalized_column_name != raw_column_name and table_empty:
-
-                self.add_column_mapping(cur, table_name, raw_column_name, canonicalized_column_name, column_schema)
-                existing_columns_raw_names.append(canonicalized_column_name)
-                self.add_column(cur,
-                                table_name,
-                                canonicalized_column_name,
-                                column_schema)
-                existing_columns[canonicalized_column_name] = column_schema
-
-            ## NEW COLUMN, VALID NAME
-            elif canonicalized_column_name != raw_column_name:
-
-                self.add_column_mapping(cur, table_name, raw_column_name, canonicalized_column_name,
-                                        nullable_column_schema)
-                existing_columns_raw_names.append(canonicalized_column_name)
-                self.add_column(cur,
-                                table_name,
-                                canonicalized_column_name,
-                                nullable_column_schema)
-                existing_columns[canonicalized_column_name] = nullable_column_schema
-
-            ## UNKNOWN
-            else:
-                raise PostgresError(
-                    'UNKNOWN: Cannot handle merging column `{}` (canonicalized as: `{}`, canonicalized with type as: `{}`) in table `{}`.'.format(
-                        self.postgres_schema,
-                        raw_column_name,
-                        canonicalized_column_name,
-                        canonicalized_typed_column_name,
-                        table_name
-                    ))
